requirements.txt
mapillary-tools
python-dotenv
streamlit
